Add AP mode and telnet server for configuration
Create an alternative mode for device configuration, the wireless
interface is put in Access Point mode and a simple telnet server is
started. When connected to the telnet server its possible to erase flash
contents, change, query and list configuration values. To enter this
configuration mode the button(GPIO15) must be pressed during startup, after
power-on or reset.